requirements.txt
py-cord~=2.0.0b1
python-dotenv
jishaku
psutil
asyncpg
DiscordUtils
pytest
aiosqlite_pool.zip
time-str
url-parser
humanize
Pillow
git+https://github.com/Rapptz/discord-ext-menus
git+https://github.com/BobDotCom/sr_api
